refactor: workspace routes
